backend/requirements.txt
torch
torchaudio
sounddevice
numpy
firebase-admin
python-dotenv
pyusb
requests
